renamed to EmptyListView, added spacing and applied to HomeView
